refactor(sql): change id type from bigint to int

+++ blog.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Host: localhost
--- Generation Time: Apr 12, 2020 at 01:39 PM
+-- Generation Time: Apr 25, 2020 at 02:21 AM
 -- Server version: 5.7.26
 -- PHP Version: 7.3.4
 
@@ -27,14 +27,15 @@ SET time_zone = "+00:00";
 --
 -- Table structure for table `article`
 --
+
 DROP TABLE IF EXISTS `article`;
 CREATE TABLE `article` (
-  `article_id` bigint(20) UNSIGNED NOT NULL COMMENT '博文ID',
-  `user_id` bigint(20) UNSIGNED NOT NULL COMMENT '发表用户ID',
+  `article_id` int(10) UNSIGNED NOT NULL COMMENT '博文ID',
+  `user_id` int(10) UNSIGNED NOT NULL COMMENT '发表用户ID',
   `article_title` tinytext NOT NULL COMMENT '博文标题',
   `article_content` longtext COMMENT '博文内容',
-  `article_views` bigint(20) UNSIGNED NOT NULL COMMENT '浏览量',
-  `article_comment_count` bigint(20) UNSIGNED NOT NULL COMMENT '评论总数',
+  `article_views` int(10) UNSIGNED NOT NULL COMMENT '浏览量',
+  `article_comment_count` int(10) UNSIGNED NOT NULL COMMENT '评论总数',
   `article_date` datetime NOT NULL COMMENT '发表时间',
   `article_last_modify_date` datetime NOT NULL COMMENT '最近一次修改时间'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -44,10 +45,11 @@ CREATE TABLE `article` (
 --
 -- Table structure for table `article_category`
 --
+
 DROP TABLE IF EXISTS `article_category`;
 CREATE TABLE `article_category` (
-  `article_id` bigint(20) UNSIGNED NOT NULL COMMENT '文章ID',
-  `category_id` bigint(20) UNSIGNED NOT NULL COMMENT '分类ID'
+  `article_id` int(10) UNSIGNED NOT NULL COMMENT '文章ID',
+  `category_id` int(10) UNSIGNED NOT NULL COMMENT '分类ID'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 -- --------------------------------------------------------
@@ -55,10 +57,11 @@ CREATE TABLE `article_category` (
 --
 -- Table structure for table `article_label`
 --
+
 DROP TABLE IF EXISTS `article_label`;
 CREATE TABLE `article_label` (
-  `article_id` bigint(20) UNSIGNED NOT NULL COMMENT '文章ID',
-  `label_id` bigint(20) UNSIGNED NOT NULL COMMENT '标签ID'
+  `article_id` int(10) UNSIGNED NOT NULL COMMENT '文章ID',
+  `label_id` int(10) UNSIGNED NOT NULL COMMENT '标签ID'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 -- --------------------------------------------------------
@@ -66,9 +69,10 @@ CREATE TABLE `article_label` (
 --
 -- Table structure for table `category`
 --
+
 DROP TABLE IF EXISTS `category`;
 CREATE TABLE `category` (
-  `category_id` bigint(20) UNSIGNED NOT NULL COMMENT '分类ID',
+  `category_id` int(10) UNSIGNED NOT NULL COMMENT '分类ID',
   `category_name` varchar(50) NOT NULL COMMENT '分类名称',
   `category_alias` varchar(15) DEFAULT NULL COMMENT '分类别名',
   `category_description` tinytext COMMENT '分类描述'
@@ -79,16 +83,17 @@ CREATE TABLE `category` (
 --
 -- Table structure for table `comment`
 --
+
 DROP TABLE IF EXISTS `comment`;
 CREATE TABLE `comment` (
-  `comment_id` bigint(20) UNSIGNED NOT NULL COMMENT '评论ID',
-  `user_id` bigint(20) UNSIGNED DEFAULT NULL COMMENT '发表用户ID',
+  `comment_id` int(10) UNSIGNED NOT NULL COMMENT '评论ID',
+  `user_id` int(10) UNSIGNED DEFAULT NULL COMMENT '发表用户ID',
   `user_name` varchar(20) DEFAULT 'Anonymity' COMMENT '发表用户名',
   `user_ip` varchar(20) NOT NULL COMMENT '用户IP',
-  `article_id` bigint(20) UNSIGNED NOT NULL COMMENT '评论博文ID',
+  `article_id` int(10) UNSIGNED NOT NULL COMMENT '评论博文ID',
   `comment_date` datetime DEFAULT NULL COMMENT '评论日期',
   `comment_content` text NOT NULL COMMENT '评论内容',
-  `parent_comment_id` bigint(20) UNSIGNED DEFAULT NULL COMMENT '父评论ID'
+  `parent_comment_id` int(10) UNSIGNED DEFAULT NULL COMMENT '父评论ID'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 -- --------------------------------------------------------
@@ -96,9 +101,10 @@ CREATE TABLE `comment` (
 --
 -- Table structure for table `label`
 --
+
 DROP TABLE IF EXISTS `label`;
 CREATE TABLE `label` (
-  `label_id` bigint(20) UNSIGNED NOT NULL COMMENT '标签ID',
+  `label_id` int(10) UNSIGNED NOT NULL COMMENT '标签ID',
   `label_name` varchar(20) NOT NULL COMMENT '标签名称',
   `label_alias` varchar(20) DEFAULT NULL COMMENT '标签别名',
   `label_description` tinytext COMMENT '标签描述'
@@ -109,9 +115,10 @@ CREATE TABLE `label` (
 --
 -- Table structure for table `option`
 --
+
 DROP TABLE IF EXISTS `option`;
 CREATE TABLE `option` (
-  `option_id` bigint(20) UNSIGNED NOT NULL COMMENT '选项ID',
+  `option_id` int(10) UNSIGNED NOT NULL COMMENT '选项ID',
   `option_name` varchar(190) NOT NULL DEFAULT '' COMMENT '选项名',
   `option_value` longtext NOT NULL COMMENT '选项值',
   `option_enable` varchar(10) NOT NULL DEFAULT 'yes' COMMENT '选项状态'
@@ -140,9 +147,10 @@ INSERT INTO `option` (`option_id`, `option_name`, `option_value`, `option_enable
 --
 -- Table structure for table `user`
 --
+
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
-  `user_id` bigint(20) UNSIGNED NOT NULL COMMENT '用户ID',
+  `user_id` int(10) UNSIGNED NOT NULL COMMENT '用户ID',
   `user_name` varchar(20) NOT NULL COMMENT '用户名',
   `user_password` varchar(255) NOT NULL COMMENT '用户密码',
   `user_email` varchar(30) DEFAULT NULL COMMENT '用户邮箱',
@@ -225,37 +233,37 @@ ALTER TABLE `user`
 -- AUTO_INCREMENT for table `article`
 --
 ALTER TABLE `article`
-  MODIFY `article_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '博文ID';
+  MODIFY `article_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '博文ID';
 
 --
 -- AUTO_INCREMENT for table `category`
 --
 ALTER TABLE `category`
-  MODIFY `category_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '分类ID';
+  MODIFY `category_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '分类ID';
 
 --
 -- AUTO_INCREMENT for table `comment`
 --
 ALTER TABLE `comment`
-  MODIFY `comment_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '评论ID';
+  MODIFY `comment_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '评论ID';
 
 --
 -- AUTO_INCREMENT for table `label`
 --
 ALTER TABLE `label`
-  MODIFY `label_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '标签ID';
+  MODIFY `label_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '标签ID';
 
 --
 -- AUTO_INCREMENT for table `option`
 --
 ALTER TABLE `option`
-  MODIFY `option_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '选项ID', AUTO_INCREMENT=13;
+  MODIFY `option_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '选项ID', AUTO_INCREMENT=13;
 
 --
 -- AUTO_INCREMENT for table `user`
 --
 ALTER TABLE `user`
-  MODIFY `user_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '用户ID';
+  MODIFY `user_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '用户ID';
 
 --
 -- Constraints for dumped tables
