+++ src/main/java/me/jinwenjie/dao/CategoryDao.java
@@ -3,13 +3,13 @@ package me.jinwenjie.dao;
 import me.jinwenjie.model.Category;
 
 public interface CategoryDao {
-    int deleteByPrimaryKey(Long categoryId);
+    int deleteByPrimaryKey(Integer categoryId);
 
     int insert(Category record);
 
     int insertSelective(Category record);
 
-    Category selectByPrimaryKey(Long categoryId);
+    Category selectByPrimaryKey(Integer categoryId);
 
     int updateByPrimaryKeySelective(Category record);
