+++ src/main/java/me/jinwenjie/model/ArticleCategoryKey.java
@@ -1,11 +1,11 @@
 package me.jinwenjie.model;
 
 public class ArticleCategoryKey {
-    private Long articleId;
+    private Integer articleId;
 
-    private Long categoryId;
+    private Integer categoryId;
 
-    public ArticleCategoryKey(Long articleId, Long categoryId) {
+    public ArticleCategoryKey(Integer articleId, Integer categoryId) {
         this.articleId = articleId;
         this.categoryId = categoryId;
     }
@@ -14,19 +14,19 @@ public class ArticleCategoryKey {
         super();
     }
 
-    public Long getArticleId() {
+    public Integer getArticleId() {
         return articleId;
     }
 
-    public void setArticleId(Long articleId) {
+    public void setArticleId(Integer articleId) {
         this.articleId = articleId;
     }
 
-    public Long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
 }
